% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADAMM11.R
\name{admm.MADMMplasso}
\alias{admm.MADMMplasso}
\title{Fit the ADMM part of  model for a given lambda vale}
\usage{
admm.MADMMplasso(
  beta0,
  theta0,
  beta,
  beta_hat,
  theta,
  rho1,
  X,
  Z,
  max_it,
  W_hat,
  XtY,
  y,
  N,
  p,
  K,
  e.abs,
  e.rel,
  alpha,
  lambda,
  alph,
  svd.w,
  tree,
  my_print = TRUE,
  invmat,
  V,
  Q,
  E,
  EE,
  O,
  P,
  H,
  HH,
  cv = cv,
  gg = 0.2
)
}
\arguments{
\item{X}{n by p matrix of predictors}

\item{Z}{n by nz matrix of modifying variables. The elements of z  may represent quantitative or categorical variables, or a mixture of the two.
Categorical varables should be coded by 0-1 dummy variables: for a k-level variable, one can use either k or k-1  dummy variables.}
}
\value{
predicted values for the ADMM part
}
\description{
Fit the ADMM part of  model for a given lambda vale
}
