% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADAMM11.R
\name{admm_MADMMplasso}
\alias{admm_MADMMplasso}
\title{Fit the ADMM part of  model for a given lambda vale}
\usage{
admm_MADMMplasso(
  beta0,
  theta0,
  beta,
  beta_hat,
  theta,
  rho1,
  X,
  Z,
  max_it,
  W_hat,
  XtY,
  y,
  N,
  e.abs,
  e.rel,
  alpha,
  lambda,
  alph,
  svd.w,
  tree,
  my_print = T,
  invmat,
  cv = cv,
  gg = 0.2,
  legacy = FALSE
)
}
\arguments{
\item{beta0}{TODO: fill paramater description}

\item{theta0}{TODO: fill paramater description}

\item{beta}{TODO: fill paramater description}

\item{beta_hat}{TODO: fill paramater description}

\item{theta}{TODO: fill paramater description}

\item{rho1}{TODO: fill paramater description}

\item{X}{n by p matrix of predictors}

\item{Z}{n by nz matrix of modifying variables. The elements of z
may represent quantitative or categorical variables, or a mixture of the two.
Categorical varables should be coded by 0-1 dummy variables: for a k-level
variable, one can use either k or k-1  dummy variables.}

\item{max_it}{TODO: fill paramater description}

\item{W_hat}{TODO: fill paramater description}

\item{XtY}{TODO: fill paramater description}

\item{y}{TODO: fill paramater description}

\item{N}{TODO: fill paramater description}

\item{e.abs}{TODO: fill paramater description}

\item{e.rel}{TODO: fill paramater description}

\item{alpha}{TODO: fill paramater description}

\item{lambda}{TODO: fill paramater description}

\item{alph}{TODO: fill paramater description}

\item{svd.w}{TODO: fill paramater description}

\item{tree}{TODO: fill paramater description}

\item{my_print}{TODO: fill paramater description}

\item{invmat}{TODO: fill paramater description}

\item{cv}{TODO: fill paramater description}

\item{gg}{TODO: fill paramater description}

\item{legacy}{If \code{TRUE}, use the R version of the algorithm. Defaults to
C++.}
}
\value{
predicted values for the ADMM part
}
\description{
TODO: add description
}
