% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADAMM11.R
\name{compute_pliable}
\alias{compute_pliable}
\title{TODO: fill this field}
\usage{
compute_pliable(X, Z, theta)
}
\arguments{
\item{X}{TODO: fill this field}

\item{Z}{TODO: fill this field}

\item{theta}{TODO: fill this field}
}
\value{
TODO: fill this field
}
\description{
TODO: fill this field
}
