% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADAMM11.R
\name{convNd2T}
\alias{convNd2T}
\title{Fit a  pliable lasso model over a path of regularization values
to fit the plasso model with ADMM}
\usage{
convNd2T(Nd, w, w_max)
}
\description{
Fit a  pliable lasso model over a path of regularization values
to fit the plasso model with ADMM
}
