% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADAMM11.R
\name{cv.MADMMplasso}
\alias{cv.MADMMplasso}
\title{Carries out cross-validation for  a  pliable lasso model over a path of regularization values}
\usage{
cv.MADMMplasso(
  fit,
  nfolds,
  X,
  Z,
  y,
  alpha = 0.5,
  lambda = fit$Lambdas,
  max_it = 50000,
  e.abs = 0.001,
  e.rel = 0.001,
  nlambda,
  rho = 5,
  my_print = F,
  alph = 1,
  foldid = NULL,
  parallel = T,
  pal = 0,
  gg = c(7, 0.5),
  TT,
  tol = 1e-04,
  cl = 2
)
}
\arguments{
\item{fit}{object returned by the pliable function}

\item{nfolds}{number of cross-validation folds
#' @param foldid  vector with values in 1:K, indicating folds for K-fold CV. Default NULL}

\item{X}{N by p matrix of predictors}

\item{Z}{N by K matrix of modifying variables. The elements of Z  may represent quantitative or categorical variables, or a mixture of the two.
Categorical varables should be coded by 0-1 dummy variables: for a k-level variable, one can use either k or k-1  dummy variables.}

\item{y}{N by D-matrix of responses. The X and Z variables are centered in the function. We recommmend that x and z also be standardized before the call}
}
\value{
predicted values
}
\description{
Carries out cross-validation for  a  pliable lasso model over a path of regularization values
}
