% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADAMM11.R
\name{generate_my_w}
\alias{generate_my_w}
\title{Generate the matrix W as used in Appendix I for use in the function.}
\usage{
generate_my_w(X = matrix(), Z = matrix(), quad = TRUE)
}
\arguments{
\item{X}{TODO: fill in description}

\item{Z}{TODO: fill in description}

\item{quad}{TODO: fill in description}
}
\description{
Generate the matrix W as used in Appendix I for use in the function.
}
