% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADAMM11.R
\name{MADMMplasso}
\alias{MADMMplasso}
\title{Fit a multi-response pliable lasso model over a path of regularization values}
\usage{
MADMMplasso(
  X,
  Z,
  y,
  alpha,
  my_lambda = NULL,
  lambda_min = 0.001,
  max_it = 50000,
  e.abs = 0.001,
  e.rel = 0.001,
  maxgrid,
  nlambda,
  rho = 5,
  my_print = F,
  alph = 1.8,
  tree,
  cv = F,
  parallel = T,
  pal = 0,
  gg = 0.05,
  tol = 1e-04,
  cl = 4
)
}
\arguments{
\item{X}{N by p matrix of predictors}

\item{Z}{N by K matrix of modifying variables. The elements of Z  may represent quantitative or categorical variables, or a mixture of the two.
Categorical varables should be coded by 0-1 dummy variables: for a k-level variable, one can use either k or k-1  dummy variables.}

\item{y}{N by D matrix  of responses. The X and Z variables are centered in the function. We recommmend that X and Z also be standardized before the call}

\item{alpha}{mixing parameter- default 0.5
#' @param max_it maximum number of iterations in the ADMM algorithm for one lambda. Default 50000}

\item{e.abs}{absolute error for the admm}

\item{e.rel}{relative error for the admm}

\item{maxgrid}{similar to nlambda}

\item{nlambda}{number of lambda values desired (default 50).}

\item{rho}{the Lagrange variable for the ADMM}

\item{gg}{penalty term for the tree structure}

\item{lambda.min.ratio}{the smallest value for lambda , as a fraction of
lambda.max, the (data derived) entry value (i.e. the smallest value for which all coefficients are zero).
Default is 0.001 if n>p, and 0.01 if n< p.}
}
\description{
Fit a multi-response pliable lasso model over a path of regularization values
}
