% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADAMM11.R
\name{plot.MADMMplasso}
\alias{plot.MADMMplasso}
\title{plot coefficients from a "MADMMplasso" object}
\usage{
\method{plot}{MADMMplasso}(x)
}
\description{
Produces a coefficient profile plot of the coefficient paths for a fitted
\code{"MADMMplasso"} object for each response.
}
