% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADAMM11.R
\name{predict.MADMMplasso}
\alias{predict.MADMMplasso}
\title{Compute predicted values from a fitted pliable  object
Make predictions from a fitted pliable lasso model}
\usage{
\method{predict}{MADMMplasso}(object, X, Z, y, lambda = NULL, ...)
}
\arguments{
\item{object}{object returned from a call to pliable}

\item{X}{N by p matrix of predictors}

\item{Z}{n by nz matrix of modifying variables. These may be observed or the predictions from a supervised learning
algorithm that predicts z from test features x  and possibly other features.}

\item{y}{N by D matrix  of responses.}

\item{lambda}{TODO: fill in description}

\item{...}{additional arguments to the generic \code{predict()} method
@return  predicted values}
}
\description{
Compute predicted values from a fitted pliable  object
Make predictions from a fitted pliable lasso model
}
