% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_parms.R
\name{tree_parms}
\alias{tree_parms}
\title{Tree parameters # TODO: check}
\usage{
tree_parms(y = y, h = 0.7)
}
\arguments{
\item{y}{N by D matrix of response variables
@param h is the tree cut off}

\item{h}{h # TODO: add description!}
}
\value{
A trained the tree with the following components:
Tree: the tree matrix stored in 1's and 0's
Tw: tree weights assocuated with the tree matrix. Each weight corresponds to a row in the tree matrix.
}
\description{
Calculate tree parameters # TODO: add description!
}
\details{
Fit the hierarchical tree structure
}
