% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADAMM11.R
\name{cv.MADMMplasso}
\alias{cv.MADMMplasso}
\title{Carries out cross-validation for  a  pliable lasso model over a path of regularization values}
\usage{
cv.MADMMplasso(
  fit,
  nfolds,
  X,
  Z,
  y,
  alpha = 0.5,
  lambda = fit$Lambdas,
  max_it = 50000,
  e.abs = 0.001,
  e.rel = 0.001,
  nlambda,
  rho = 5,
  my_print = F,
  alph = 1,
  foldid = NULL,
  parallel = T,
  pal = 0,
  gg = c(7, 0.5),
  TT,
  tol = 1e-04,
  cl = 2
)
}
\arguments{
\item{fit}{object returned by the pliable function}

\item{nfolds}{number of cross-validation folds}

\item{X}{N by p matrix of predictors}

\item{Z}{N by K matrix of modifying variables. The elements of Z  may
represent quantitative or categorical variables, or a mixture of the two.
Categorical varables should be coded by 0-1 dummy variables: for a k-level
variable, one can use either k or k-1  dummy variables.}

\item{y}{N by D-matrix of responses. The X and Z variables are centered in
the function. We recommmend that x and z also be standardized before the call}

\item{alpha}{TODO: add parameter description}

\item{lambda}{TODO: add parameter description}

\item{max_it}{TODO: add parameter description}

\item{e.abs}{TODO: add parameter description}

\item{e.rel}{TODO: add parameter description}

\item{nlambda}{TODO: add parameter description}

\item{rho}{TODO: add parameter description}

\item{my_print}{TODO: add parameter description}

\item{alph}{TODO: add parameter description}

\item{foldid}{TODO: add parameter description}

\item{parallel}{TODO: add parameter description}

\item{pal}{TODO: add parameter description}

\item{gg}{TODO: add parameter description}

\item{TT}{TODO: add parameter description}

\item{tol}{TODO: add parameter description}

\item{cl}{TODO: add parameter description}
}
\value{
results containing the CV values
}
\description{
Carries out cross-validation for  a  pliable lasso model over a path of regularization values
}
