% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim2.R
\name{sim2}
\alias{sim2}
\title{Simulate data for the model. This is the second simulation data used in the paper}
\usage{
sim2(p = 500, n = 100, m = 24, nz = 4, rho = 0.4, B.elem = 0.5)
}
\arguments{
\item{p}{column for X which is the main effect}

\item{n}{number of observations}

\item{m}{number of responses}

\item{nz}{number of modifiers}

\item{rho}{values to be used in the covariance matrix when generating  X}

\item{B.elem}{the value of the non-zero elements in beta?}
}
\value{
The simulated data with the following components:
Beta: matrix of actual beta coefficients  p by m
Theta: a m by p by K array of actual theta coefficients
Y: a N by m matrix of response variables
X: a N by p matrix of covariates
Z: a N by K matrix of modifiers
}
\description{
Simulate data for the model
}
