% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADAMM11.R
\name{tree.parms}
\alias{tree.parms}
\title{Fit the hierarchical tree structure}
\usage{
tree.parms(y = y, h = 0.7)
}
\arguments{
\item{y}{N by D matrix of response variables
@param h is the tree cut off}
}
\value{
A trained the tree with the following components:
Tree: the tree matrix stored in 1's and 0's
Tw: tree weights assocuated with the tree matrix. Each weight corresponds to a row in the tree matrix.
}
\description{
Fit the hierarchical tree structure
}
