% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_parms.R
\name{tree_parms}
\alias{tree_parms}
\title{Fit the hierarchical tree structure}
\usage{
tree_parms(y = y, h = 0.7)
}
\arguments{
\item{y}{N by D matrix of response variables}

\item{h}{is the tree cut off}
}
\value{
A trained  tree with the following components:
Tree: the tree matrix stored in 1s and 0s
Tw: tree weights associated with the tree matrix. Each weight corresponds to a row in the tree matrix.
h_clust: Summary of the hclust call
y.colnames: names of the response
}
\description{
Fit the hierarchical tree structure
}
